+++ packages/engine/src/map/index.ts
@@ -1,9 +1,8 @@
 import * as THREE from 'three'
 import { MapboxTileLoader } from './MapboxTileLoader'
-import { fetchTiles, buildMesh } from './MeshBuilder'
+import { buildMesh } from './MeshBuilder'
+import { fetchTiles } from './MapBoxClient'
 import { MapProps } from './MapProps'
-
-type ILayerName = 'building' | 'road'
 
 const useNew = true
