+++ packages/engine/src/map/types.ts
@@ -1,2 +1,8 @@
 
+import { Feature } from 'geojson'
+
 export type ILayerName = 'building' | 'road'
+export interface TileFeaturesByLayer {
+  building: Feature[]
+  road: Feature[]
+}
